+++ desmume/src/bios.c
@@ -993,7 +993,7 @@ u32 Diff16bitUnFilter(armcpu_t* cpu)
 
 u32 bios_sqrt(armcpu_t* cpu)
 {
-     cpu->R[0] = (u32)sqrt((double)(cpu->R[0]));
+     cpu->R[0] = (u32)sqrt((float)(cpu->R[0]));
      return 1;
 }
